- Added cumulus record creation.
